fix: address cold eyes review issues in data collation and attention
- Use tokenizer.pad_token_id instead of hard-coded 1 for ntokens count
- Add float32 upcast in two-stream attention softmax (bf16/fp16 stability)
- Handle empty word boundary list by falling back to token-level permutation
- Add bounds check assertion before Cython span permutation call
- Use max(1, round()) for pred_size to avoid masking entire short sequences
- Always seed collator RNGs for deterministic resume (time-based if unspecified)

+++ annotated_mpnet/modeling/mpnet_for_pretraining.py
@@ -539,7 +539,10 @@ def two_stream_self_attention(
                 attn_weights = fill_mask(attn_weights, mask)
 
             # Softmax the energy to get the final attention weights
-            attn_weights = F.softmax(attn_weights, dim=-1).type_as(attn_weights)
+            # Upcast to float32 before softmax to avoid bf16/fp16 overflow, matching standard attention
+            attn_weights = F.softmax(attn_weights, dim=-1, dtype=torch.float32).type_as(
+                attn_weights
+            )
 
             # Do the attention dropout
             attn_weights = F.dropout(attn_weights, p=self.dropout, training=self.training)
